requirements.txt
discord.py>=2.3.0
claude-agent-sdk>=0.1.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
colorama>=0.4.6
flake8>=6.0.0
